backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Redis
redis[hiredis]==5.1.1

# Telegram
python-telegram-bot==21.6

# AI
openai==1.54.0

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4


# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Utils
httpx==0.27.2
python-multipart==0.0.12
pillow==11.0.0
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
av==13.1.0
apscheduler==3.10.4

